+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ## v0.2.0
 - Add classes IDTokenVerifier and AccessTokenVerifier
+- Add DeprecationWarning to not use JWTVerifier class
+- Add proxy support
 - Update README
 - Few codebase improvements
 
 _New features:_
-Separate classes for verifying ID Tokens and Access Tokens
+- Separate classes for verifying ID Tokens and Access Tokens
+- Add proxy support
 
 ## v0.1.0
 - Initial release